Move utils into root
