+++ src/MCpypack/recipe/crafting_decorated_pot.py
@@ -3,7 +3,7 @@
 from packaging.version import Version
 
 from MCpypack.item import ItemLike
-from .utils import SimpleResult, CountedResult
+from MCpypack.utils import CountedResult
 from .recipe import Recipe
 
 class CraftingDecoratedPot(Recipe):
