+++ src/MCpypack/recipe/smithing.py
@@ -8,7 +8,7 @@ from .recipe import Recipe
 from MCpypack.item import ItemLike
 from MCpypack.item import trim_pattern
 TrimPattern = trim_pattern.TrimPattern
-from .utils import SimpleResult
+from MCpypack.utils import SimpleResult
 
 class SmithingTransform(Recipe):
     """
