+++ src/MCpypack/utils/result.py
